Fixes startup issues

+++ src/Models/PagedResults.cs
@@ -10,6 +10,6 @@ namespace MessianicChords.Models
         public int Skip { get; init; }
         public int Take { get; init; }
         public List<T> Results { get; init; } = new List<T>();
-        public int TotalCount { get; init; }
+        public long TotalCount { get; init; }
     }
 }
